Add transaction.rs
